+++ src/views/HomeView.vue
@@ -3,12 +3,10 @@
     <h1>This is the Home View</h1>
     <p>Content coming soon...</p>
     <Card/>
-    <MapsRoutes/>
   </div>
 </template>
 
 <script setup>
 import Card from '../components/ui/card.vue';
-import MapsRoutes from '../components/ui/MapsRoutes.vue'
 // Component logic goes here
 </script>
